requirements.txt
# requirements.txt

# Core dependencies
boto3==1.26.137
requests==2.28.2
python-dotenv==0.21.1
Pillow==9.5.0

# Microsoft Graph integration
msal==1.20.0

# Database
SQLAlchemy==1.4.41

# Testing
pytest==7.3.1
pytest-cov==4.0.0
pytest-asyncio==0.21.0
pytest-timeout==2.1.0

# Development
black==23.3.0
flake8==6.0.0
mypy==1.3.0

selenium==4.8.0